Added logo, re-structured a bit.
