+++ public/manifest.json
@@ -2,9 +2,9 @@
   "name": "TBD",
   "short_name": "TBD",
   "start_url": "TBD",
-  "display": "TBD",
-  "theme_color": "TBD",
-  "background_color": "TBD",
+  "display": "",
+  "theme_color": "#FFFFFF",
+  "background_color": "#FFFFFF",
   "description": "TBD",
   "icons": [],
   "related_applications": []
